Youtube_Summary/requirements.txt
streamlit
python-dotenv
youtube-transcript-api
langchain
langchain-openai
tiktoken
urllib3